refactor(mobile-target-claim): rename files to match convention
